Refactor: mainpageInterface 분리관리
