+++ src/hooks/query/usePopularCourses.ts
@@ -1,22 +1,6 @@
 import { useSimpleQuery } from '../../api/Helper/useSimpleQuery'
 import { api } from '../../api/client'
-
-export interface Lecture {
-  id: number
-  uuid: string
-  title: string
-  instructor: string
-  thumbnail_img_url: string
-  categories: { id: number; name: string }[]
-  difficulty: string
-  original_price: number
-  discount_price: number
-  platform: string
-  average_rating: number
-  duration: number
-  url_link: string
-  is_bookmarked: boolean
-}
+import type { Lecture } from '../../types/apiInterface/mainpageInterface.ts'
 
 const fetchPopularCourses = async (): Promise<Lecture[]> => {
   const response = await api.get<{
